refactor(scim): extract user-flow sync + patch + helpers (#149 slice 2)
Second slice of the scim/users.go decomposition.
Lifts the per-RPC support functions out of users.go into a sibling
users_helpers.go file:

- newULID (free function, top-of-file)
- syncUserFromSCIM + syncIdentityLink (sync helpers)
- extractNameFromPatchOps + formatExternalName (patch-op extractors)
- verifyProviderOwnership (provider-ownership guard)
- baseURLFromRequest (request-shape helper)
- linuxUsernameSanitizeRe + deriveLinuxUsername (Linux-username derivation)

users.go: 1106 \xe2\x86\x92 897 LOC (-209). Cumulative since baseline 1273: -376.

users.go now contains only the per-RPC HTTP handlers (list /
listFiltered / get / create / replace / patch / handleUserPatchReplace
/ delete). Pure file-shuffling \xe2\x80\x94 zero behaviour change.

Stacked on #231 (slice 1) \xe2\x80\x94 will rebase as it lands.

All SCIM v2 conformance tests still pass (206s integration suite).

Refs manchtools/power-manage-server#149 (audit F009).
