+++ internal/scim/users.go
@@ -2,28 +2,22 @@ package scim
 
 import (
 	"context"
-	"crypto/rand"
 	"encoding/json"
 	"errors"
 	"fmt"
 	"net/http"
-	"regexp"
 	"strconv"
 	"strings"
-	"time"
 
 	"github.com/jackc/pgx/v5"
-	"github.com/oklog/ulid/v2"
 
 	"github.com/manchtools/power-manage/server/internal/eventtypes"
 	"github.com/manchtools/power-manage/server/internal/store"
 	db "github.com/manchtools/power-manage/server/internal/store/generated"
 )
 
-func newULID() string {
-	entropy := ulid.Monotonic(rand.Reader, 0)
-	return ulid.MustNew(ulid.Timestamp(time.Now()), entropy).String()
-}
+// newULID + Linux-username derivation + sync/patch helpers all live
+// in users_helpers.go. SCIM-resource shapers live in users_translation.go.
 
 // listUsers handles GET /scim/v2/{slug}/Users
 func (h *Handler) listUsers(w http.ResponseWriter, r *http.Request) {
@@ -897,210 +891,7 @@ func (h *Handler) deleteUser(w http.ResponseWriter, r *http.Request) {
 // SCIM user-resource shapers (userToSCIM, userRowToSCIM, etc.) +
 // safeNameField live in users_translation.go.
 
-// syncUserFromSCIM syncs email, active status, profile, and identity link data from SCIM.
-// SCIM is treated as the source of truth — any differences are overwritten.
-func (h *Handler) syncUserFromSCIM(ctx context.Context, provider db.IdentityProvidersProjection, userID, email string, active *bool, name *SCIMName) {
-	user, err := h.store.Queries().GetUserByID(ctx, userID)
-	if err != nil {
-		h.logger.Error("failed to get user for SCIM sync", "user_id", userID, "error", err)
-		return
-	}
-
-	// Sync email
-	if email != "" && email != user.Email {
-		h.appendEvent(ctx, store.Event{
-			StreamType: "user",
-			StreamID:   userID,
-			EventType:  string(eventtypes.UserEmailChanged),
-			Data:       map[string]any{"email": email},
-			ActorType:  "scim",
-			ActorID:    provider.ID,
-		})
-	}
-
-	// Sync active status (nil = not provided, default to true per SCIM RFC 7643)
-	isActive := active == nil || *active
-	if !isActive && !user.Disabled {
-		h.appendEvent(ctx, store.Event{
-			StreamType: "user",
-			StreamID:   userID,
-			EventType:  string(eventtypes.UserDisabled),
-			Data:       map[string]any{},
-			ActorType:  "scim",
-			ActorID:    provider.ID,
-		})
-	} else if isActive && user.Disabled {
-		h.appendEvent(ctx, store.Event{
-			StreamType: "user",
-			StreamID:   userID,
-			EventType:  string(eventtypes.UserEnabled),
-			Data:       map[string]any{},
-			ActorType:  "scim",
-			ActorID:    provider.ID,
-		})
-	}
-
-	// Sync profile fields (display_name, given_name, family_name)
-	newDisplayName := formatExternalName(name)
-	newGivenName := safeNameField(name, "given")
-	newFamilyName := safeNameField(name, "family")
-	if newDisplayName != "" || newGivenName != "" || newFamilyName != "" {
-		h.appendEvent(ctx, store.Event{
-			StreamType: "user",
-			StreamID:   userID,
-			EventType:  string(eventtypes.UserProfileUpdated),
-			Data: map[string]any{
-				"display_name": newDisplayName,
-				"given_name":   newGivenName,
-				"family_name":  newFamilyName,
-			},
-			ActorType: "scim",
-			ActorID:   provider.ID,
-		})
-	}
-
-	// Sync identity link (external_email + external_name)
-	h.syncIdentityLink(ctx, provider, userID, email, name)
-}
-
-// syncIdentityLink updates the identity link's external_email and external_name
-// to reflect the latest data from the SCIM provider (source of truth).
-func (h *Handler) syncIdentityLink(ctx context.Context, provider db.IdentityProvidersProjection, userID, email string, name *SCIMName) {
-	link, err := h.store.Queries().GetIdentityLinkByProviderAndUser(ctx, db.GetIdentityLinkByProviderAndUserParams{
-		ProviderID: provider.ID,
-		UserID:     userID,
-	})
-	if err != nil {
-		h.logger.Error("failed to get identity link for SCIM sync", "user_id", userID, "provider_id", provider.ID, "error", err)
-		return
-	}
-
-	h.appendEvent(ctx, store.Event{
-		StreamType: "identity_provider",
-		StreamID:   link.ID,
-		EventType:  string(eventtypes.IdentityLinkLoginUpdated),
-		Data: map[string]any{
-			"provider_id":    provider.ID,
-			"external_id":    link.ExternalID,
-			"external_email": email,
-			"external_name":  formatExternalName(name),
-		},
-		ActorType: "scim",
-		ActorID:   provider.ID,
-	})
-}
-
-// extractNameFromPatchOps scans PATCH operations for name-related changes and
-// returns a SCIMName if any were found, or nil if none.
-func extractNameFromPatchOps(ops []SCIMPatchOp) *SCIMName {
-	var name SCIMName
-	found := false
-
-	for _, op := range ops {
-		if op.Op.Normalize() != SCIMPatchOpReplace {
-			continue
-		}
-		path := strings.ToLower(op.Path)
-
-		switch path {
-		case "name":
-			// Value is a name object
-			if m, ok := op.Value.(map[string]any); ok {
-				if v, ok := m["givenName"].(string); ok {
-					name.GivenName = v
-					found = true
-				}
-				if v, ok := m["familyName"].(string); ok {
-					name.FamilyName = v
-					found = true
-				}
-				if v, ok := m["formatted"].(string); ok {
-					name.Formatted = v
-					found = true
-				}
-			}
-		case "name.givenname":
-			if v, ok := op.Value.(string); ok {
-				name.GivenName = v
-				found = true
-			}
-		case "name.familyname":
-			if v, ok := op.Value.(string); ok {
-				name.FamilyName = v
-				found = true
-			}
-		case "name.formatted":
-			if v, ok := op.Value.(string); ok {
-				name.Formatted = v
-				found = true
-			}
-		}
-	}
-
-	if !found {
-		return nil
-	}
-	return &name
-}
-
-// formatExternalName extracts a display name from SCIM name fields.
-func formatExternalName(name *SCIMName) string {
-	if name == nil {
-		return ""
-	}
-	if name.Formatted != "" {
-		return name.Formatted
-	}
-	parts := []string{}
-	if name.GivenName != "" {
-		parts = append(parts, name.GivenName)
-	}
-	if name.FamilyName != "" {
-		parts = append(parts, name.FamilyName)
-	}
-	return strings.Join(parts, " ")
-}
-
-// verifyProviderOwnership checks that the user has an identity link to the
-// given SCIM provider. This prevents one provider from accessing or modifying
-// users provisioned by a different provider.
-func (h *Handler) verifyProviderOwnership(ctx context.Context, providerID, userID string) error {
-	_, err := h.store.Queries().GetIdentityLinkByProviderAndUser(ctx, db.GetIdentityLinkByProviderAndUserParams{
-		ProviderID: providerID,
-		UserID:     userID,
-	})
-	return err
-}
-
-// baseURLFromRequest constructs the SCIM base URL from the request.
-func baseURLFromRequest(r *http.Request, slug string) string {
-	scheme := "https"
-	if r.TLS == nil {
-		if fwd := r.Header.Get("X-Forwarded-Proto"); fwd == "https" || fwd == "http" {
-			scheme = fwd
-		} else {
-			scheme = "http"
-		}
-	}
-	return fmt.Sprintf("%s://%s/scim/v2/%s", scheme, r.Host, slug)
-}
-
-var linuxUsernameSanitizeRe = regexp.MustCompile(`[^a-z0-9_.\-]`)
-
-func deriveLinuxUsername(email, preferredUsername string) string {
-	var username string
-	switch {
-	case preferredUsername != "":
-		username = preferredUsername
-	case strings.Contains(email, "@"):
-		username = email[:strings.Index(email, "@")]
-	default:
-		username = email
-	}
-	username = strings.ToLower(username)
-	username = linuxUsernameSanitizeRe.ReplaceAllString(username, "_")
-	if len(username) > 32 {
-		username = username[:32]
-	}
-	return username
-}
+// SCIM user sync helpers (syncUserFromSCIM, syncIdentityLink),
+// patch-op extractors (extractNameFromPatchOps, formatExternalName),
+// provider-ownership guard, request-shape helpers, and Linux-username
+// derivation live in users_helpers.go.
